Refactor editor components: Replace ComponentsTab with ComponentsList and remove unused loading components
